refactor: move cloud-init fixtures to conftest.py with TODO for copy-offload folder
Made-with: Cursor
